Trivial changes * Changed indenting * Renamed variables
Realization: child vs adult referred to two semantic environments, yielding different network growth values for all the same words. The previous analysis did not include a comparison between groups or participants. In the planned analysis, we are comparing autistic and nonautistic children. This is a fundamentally different kind of analysis.